Load block form styles correctly Remove the view block and move it to pro
