+++ js/src/blocks.js
@@ -4,5 +4,4 @@
  */
 
 import './form/block';
-import './view/block';
