+++ js/src/common/components/icon.js
